+++ Assets/Scripts/Simulation/ClientSimulationOtherPlayers.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 using Zenject;
 
 public class ClientSimulationOtherPlayers : ClientSimulationEntity {
-    public class Factory : PlaceholderFactory<long,ClientSimulationOtherPlayers> {}
+    public class Factory : PlaceholderFactory<int,ClientSimulationOtherPlayers> {}
 }
